backend/main.py
import os
import pandas as pd
import numpy as np
from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, List
from datetime import datetime

# Import project's prediction modules
from src.prediction.predict_payment import PaymentPredictor
from src.prediction.forecast_cashflow import CashFlowForecaster

app = FastAPI(title="CashCult ML Backend")

# Enable CORS for the Next.js frontend
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# --- Path Configuration ---
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
MODEL_DIR = os.path.join(BASE_DIR, "models")
DATA_DIR = os.path.join(BASE_DIR, "data")

# --- Initialize Predictors ---
try:
    predictor = PaymentPredictor(models_dir=MODEL_DIR, data_dir=DATA_DIR)
    forecaster = CashFlowForecaster(models_dir=MODEL_DIR)
    print("ML Models and Predictors initialized successfully.")
except Exception as e:
    print(f"Error initializing ML predictors: {e}")
    predictor = None
    forecaster = None

# --- Data Models ---
class AnalysisResponse(BaseModel):
    invoice_id: Optional[str] = None
    amount: float
    client_name: str
    issue_date: str
    due_date: str
    risk_score: float
    probability_on_time: float
    insight_text: str
    severity: str

@app.get("/")
def read_root():
    return {"status": "CashCult ML Backend Running", "models_loaded": predictor is not None}

import pdfplumber
import re
from dateutil import parser

# --- Extraction Helpers ---
def extract_text_from_pdf(file_path):
    text = ""
    with pdfplumber.open(file_path) as pdf:
        for page in pdf.pages:
            text += page.extract_text() + "\n"
    return text

def decipher_hashed_data(data_string: str):
    """
    Handles formats like '2026#03#12' or 'Invoice#101'
    """
    if isinstance(data_string, str):
        return data_string.replace("#", "-")
    return data_string

def find_regex(text, patterns):
    for pattern in patterns:
        match = re.search(pattern, text, re.IGNORECASE)
        if match:
            return match.group(1).strip()
    return None

@app.post("/analyze-invoice", response_model=AnalysisResponse)
async def analyze_invoice(file: UploadFile = File(...)):
    """
    Receives an invoice, parses PDF text, deciphers data, and runs ML inference.
    """
    try:
        # Save temp file
        temp_path = f"temp_{file.filename}"
        with open(temp_path, "wb") as buffer:
            buffer.write(await file.read())

        text = ""
        if file.filename.endswith(".pdf"):
            text = extract_text_from_pdf(temp_path)
        
        os.remove(temp_path) # Cleanup

        # --- Real-Time Extraction Logic ---
        # Decipher hashtags in the entire text first to make regex easier
        processed_text = decipher_hashed_data(text)
        
        # Regex for common fields
        invoice_id = find_regex(processed_text, [r"Invoice\s*(?:ID|#|Number)?\s*[:#-]?\s*([A-Za-z0-9-]+)"])
        client_name = find_regex(processed_text, [r"(?:Bill To|Client|Customer)\s*:?\s*([A-Za-z0-9\s]+)"])
        amount_str = find_regex(processed_text, [r"(?:Total|Amount|Sum)\s*:?\s*(?:₹|Rs\.?|INR)?\s*([\d,]+\.?\d*)"])
        
        # Date regex (looks for YYYY-MM-DD or DD-MM-YYYY)
        dates = re.findall(r"(\d{4}[-/]\d{2}[-/]\d{2}|\d{2}[-/]\d{2}[-/]\d{4})", processed_text)
        
        # Fallbacks if extraction fails
        extracted_amount = float(amount_str.replace(",", "")) if amount_str else 750000.0
        final_client_name = client_name if client_name else "Client_B_1"
        final_issue_date = dates[0] if len(dates) > 0 else datetime.now().strftime("%Y-%m-%d")
        final_due_date = dates[1] if len(dates) > 1 else (datetime.now() + timedelta(days=30)).strftime("%Y-%m-%d")
        
        # Map client name to ID for the model (assuming mapping exists in data/clients.csv)
        client_id = "C001" 
        if predictor and predictor.feature_df is not None and "client_name" in predictor.feature_df.columns:
            # Simple lookup
            match = predictor.feature_df[predictor.feature_df["client_name"].str.contains(final_client_name, case=False, na=False)]
            if not match.empty:
                client_id = match.iloc[0]["client_id"]

        if predictor:
            # Run REAL Inference
            res = predictor.predict(
                client_id=client_id,
                invoice_amount=extracted_amount,
                invoice_issue_date=final_issue_date,
                invoice_due_date=final_due_date
            )
            
            risk_score = res["client_risk_score"]
            prob_on_time = 1.0 - res["delay_probability"]
            
            # Match severity to Convex Schema: notifications expect ["critical", "warning", "info"]
            # Insights expect ["critical", "warning", "safe"]
            if res["priority_level"] == "CRITICAL":
                severity = "critical"
            elif res["priority_level"] == "HIGH":
                severity = "warning"
            else:
                severity = "safe" # Insights will use this
            
            insight = f"AI Extraction: {final_client_name} (Inv: {invoice_id or 'N/A'}). Priority: {res['priority_level']}. "
            if res["predicted_delay_days"] > 0:
                insight += f"ML predicts delay of {res['predicted_delay_days']} days."
            else:
                insight += f"Clean payment history detected."
        else:
            risk_score = 0.5
            prob_on_time = 0.5
            severity = "warning"
            insight = "ML Predictor not loaded. Using basic extraction."

        return AnalysisResponse(
            invoice_id=invoice_id,
            amount=extracted_amount,
            client_name=final_client_name,
            issue_date=final_issue_date,
            due_date=final_due_date,
            risk_score=risk_score,
            probability_on_time=prob_on_time,
            severity=severity,
            insight_text=insight
        )
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/forecast")
async def get_forecast():
    """
    Returns the real forecast from Prophet models.
    """
    if not forecaster:
        # Static fallback if models aren't ready
        return {
            "dates": ["Mar 9", "Mar 16", "Mar 23", "Mar 30", "Apr 6"],
            "p50_likely": [1200000, 1100000, 900000, 850000, 1500000],
            "p25_risk": [1000000, 800000, 600000, 500000, 1200000],
            "p75_optimistic": [1400000, 1300000, 1100000, 1050000, 1800000]
        }
    
    try:
        res = forecaster.get_forecast(days=30)
        # Transform into a format the graph expects (arrays of values)
        # The frontend current matches by index to the 'dates' array
        return {
            "dates": [str(d)[:10] for d in res["dates"]], # Format as YYYY-MM-DD
            "p50_likely": res["predicted_balance"].tolist(),
            "p25_risk": (res["predicted_balance"] * 0.8).tolist(), # Using some margin for visual P25/P75 if not present
            "p75_optimistic": (res["predicted_balance"] * 1.2).tolist()
        }
    except Exception as e:
        print(f"Forecast Error: {e}")
        return {"error": str(e)}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
