backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-multipart
pydantic
prophet
python-dotenv
xgboost
matplotlib
seaborn
pdfplumber
python-dateutil
